requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-generativeai==0.3.2
requests==2.31.0
aiohttp==3.9.3 